backend/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.32.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
crewai==0.80.0
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
python-multipart==0.0.12
aiofiles==24.1.0
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
python-docx==1.1.2